Extract sorting to lib/views/table
